wandboard: Fix SDHC platform data
Add 'always_present = 1' setting to the platform data for the ESDHC
controller which is dedicated to the Broadcom wifi device (BCM4329).
Without this change, the kernel has problems downloading firmware to
the chip.

Made some other modifications to eliminate annoying boot messages,
and the addition of the .cd_type to the data which tells the driver
what type of CD signal to expect (i.e. for permanent on-board chips
this is different from removable devices).